requirements.txt
aiogram==3.10.0
web3==6.15.1
python-dotenv==1.0.1